refactoring password generator into re-usable package
